Fixed the saving user's session logical error
Fixed by changing the logic from inserting the user's data after signup is successful in the Firebase Authentication. In addition to removed any insertion logic inside the `LoginActivity`, it is not needed at all. In the `ProfileFragment` I fetched from the Room Database a User that has the `FirebaseAuthentication.getCurrentUser().getEmail()`, and if it's not null then set the email in the profile, else create and insert a new user (this case only occurs for legacy Database's versions)